tag-generator/app/pyproject.toml
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests", "."]
addopts = [
    "--cov=.",
    "--cov-report=xml",
    "--cov-report=term-missing",
    "--junit-xml=tests/results.xml",
    "-v",
]

[project]
name = "tag-generator"
version = "0.1.0"
description = "Ultra-minimal tag generator (external models required)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Japanese Text Processing (self-contained)
    "fugashi[unidic-lite]>=1.5.1", # includes unidic-lite, no external MecAB needed

    # Language Processing (minimal)
    "nltk>=3.9.1",
    "langdetect>=1.0.9",

    # Database
    "psycopg2-binary>=2.9.0",
    "psycopg[binary]>=3.2.0",
    "psycopg-pool>=3.2.0",

    # Logging and Data Validation
    "structlog>=25.4.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # Security
    "nh3>=0.2.18",

    # Authentication and API
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "aiohttp>=3.9.0",
    "redis>=5.0.0",  # Redis Streams consumer
    "PyJWT>=2.8.0",
    "python-multipart>=0.0.6",

    # OpenTelemetry
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp-proto-http>=1.29.0",
    "opentelemetry-instrumentation-logging>=0.50b0",

    # NOTE: Heavy ML packages removed for ultra-minimal build
    # keybert, sentence-transformers, transformers must be installed in external models
    # or use runtime installation with lazy loading
]

[dependency-groups]
dev = [
    # Testing Framework
    "pytest>=8.4.1",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.1",
    "pytest-timeout>=2.4.0",

    # Code Quality Tools
    "ruff>=0.12.1",
    "bandit>=1.8.0",
    "pytest-asyncio>=0.24.0",

    # Property-based testing
    "hypothesis>=6.100.0",

    # Type Checking
    "pyrefly>=0.42.0",
    "types-psycopg2>=2.9.21.20250516",

    # Additional test dependencies for CI
    "langdetect>=1.0.9",
    "psutil>=5.9.0",
]

# ML dependencies for production and testing
# These are installed separately in Dockerfile to control installation order
ml = [
    "sentence-transformers>=3.3.0",
    "keybert>=0.8.5",
    "transformers>=4.40.0",
    "torch>=2.1.0",
    "scikit-learn>=1.5.2",
    "numpy>=1.26.0",
    "onnxruntime>=1.21.0",
]

# Extended ML dependencies including GiNZA for improved Japanese NLP
# Install with: uv sync --group ml-extended
# GiNZA provides better noun phrase extraction and named entity recognition
ml-extended = [
    # Include all base ML dependencies
    "sentence-transformers>=3.3.0",
    "keybert>=0.8.5",
    "transformers>=4.40.0",
    "torch>=2.1.0",
    "scikit-learn>=1.5.2",
    "numpy>=1.26.0",
    "onnxruntime>=1.21.0",
    # GiNZA: Japanese NLP library built on spaCy
    # Provides: noun phrase chunking, dependency parsing, NER
    "ginza>=5.2.0",
    "ja-ginza>=5.2.0",
    "spacy>=3.8.0",
]

# NOTE: ML dependencies are in a separate group to allow flexible installation
# In Dockerfile: Install PyTorch from custom index first, then other ML deps
# For development: uv sync --group ml

[tool.uv]
constraint-dependencies = ["protobuf>=6.33.5"]

[tool.setuptools.packages]
find = { include = [
    "tag_generator",
    "tag_generator.*",
    "article_fetcher",
    "article_fetcher.*",
    "tag_extractor",
    "tag_extractor.*",
    "tag_inserter",
    "tag_inserter.*",
] }

# Ultra-minimal optimization notes:
# - Removed ALL heavy ML packages from core dependencies
# - Created separate 'ml' group for ML dependencies
# - Reduced to absolute minimum for basic functionality
# - ML models must be provided externally or installed at runtime

[tool.ruff]
target-version = "py313"
line-length = 120
exclude = [".git", "__pycache__", ".venv", "build", "dist", "*.egg-info"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "S",  # bandit (security)
]
ignore = [
    "S101", # Use of assert detected (OK in tests)
    "S608", # Possible SQL injection (handled by ORM)
    "E501", # Line too long (handled by formatter when possible)
    "S110", # try-except-pass (sometimes needed in cleanup code)
]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = ["S101", "S105", "S603", "S110"]
"tests/**" = ["S101", "S105", "S603", "S110"]
"auth_service.py" = [
    "S104",
] # Allow binding to all interfaces for development server

[tool.bandit]
exclude_dirs = ["tests", ".venv", "__pycache__", "*.egg-info", "build", "dist"]
skips = [
    "B101",
    "B404",
    "B603",
    "B110",
    "B104",
    "B108",
    "B602",
] # Skip assert_used, subprocess import, subprocess execution, try_except_pass, hardcoded_bind_all_interfaces, hardcoded_tmp_directory, and subprocess_popen_with_shell_equals_true tests

[tool.pyrefly]
project-includes = ["."]
project-excludes = [
    "**/__pycache__",
    "**/.venv",
    "**/build",
    "**/dist",
    "**/*.egg-info",
]
python-version = "3.13"

# Suppress errors for optional ML dependencies (installed separately in Docker)
[tool.pyrefly.errors]
missing-import = false
missing-module-attribute = false
